[refactor] Moved BigAnimatedCounter to animatedCounter.dart
